+++ src/bgp/message/update.rs
@@ -27,6 +27,7 @@ const COFF: usize = 19; // XXX replace this with .skip()'s?
 
 /// BGP UPDATE message, variant of the [`Message`] enum.
 #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub struct UpdateMessage<Octs: Octets> {
     octets: Octs,
     session_config: SessionConfig,
@@ -521,6 +522,7 @@ impl<Octs: Octets> UpdateMessage<Octs> {
 ///
 #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub struct SessionConfig {
     pub four_octet_asn: FourOctetAsn,
     pub add_path: AddPath,
@@ -590,6 +592,7 @@ impl SessionConfig {
 /// Indicates whether this session is Four Octet capable.
 #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub enum FourOctetAsn {
     Enabled,
     Disabled,
@@ -598,6 +601,7 @@ pub enum FourOctetAsn {
 /// Indicates whether AddPath is enabled for this session.
 #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub enum AddPath {
     Enabled,
     Disabled,
@@ -810,8 +814,8 @@ impl<'a> PathAttribute<'a, [u8]> {
                 }
             },
             PathAttributeType::Aggregator => {
-                let mut pp = parser.parse_parser(len)?;
-                Aggregator::check(&mut pp, config)?;
+                let pp = parser.parse_parser(len)?;
+                Aggregator::check(&pp, config)?;
             },
             PathAttributeType::Communities => {
                 let mut pp = parser.parse_parser(len)?;
@@ -1265,6 +1269,7 @@ impl<'a, Ref: Octets> IntoIterator for PathAttributes<'a, Ref> {
 //--- Aggregator -------------------------------------------------------------
 /// Path Attribute (7).
 #[derive(Debug, Eq, PartialEq, Copy, Clone, Hash)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub struct Aggregator {
     asn: Asn,
     speaker: Ipv4Addr,
@@ -1390,7 +1395,7 @@ impl<Octs: Octets> Iterator for LargeCommunityIter<Octs> {
 }
 
 impl Aggregator {
-    fn check(parser: &mut Parser<[u8]>, config: SessionConfig)
+    fn check(parser: &Parser<[u8]>, config: SessionConfig)
         -> Result<(), ParseError>
     {
         let len = parser.remaining(); // XXX is this always correct?